Fix #284: use static_cast to fix conversion error

+++ Hazel/src/Hazel/Scene/Entity.h
@@ -40,7 +40,7 @@ namespace Hazel {
 			m_Scene->m_Registry.remove<T>(m_EntityHandle);
 		}
 
-		operator bool() const { return m_EntityHandle != 0; }
+		operator bool() const { return static_cast<int>(m_EntityHandle) != 0; }
 	private:
 		entt::entity m_EntityHandle{ 0 };
 		Scene* m_Scene = nullptr;
